Added a context menu on right click.
Enable to force a result in a the state "OK". This is usefull when a build fails because of the continous build system (the build is sane but the C.I. is malfunctionning).
The state is reported "OK". This is different from the state reported by Jenkins ("SUCCESS") so the report shows that the result was forced.